remove checker
